remove reading arg as name
