+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/liip/sheriff v0.0.0-20190308094614-91aa83a45a3d
 	github.com/sirupsen/logrus v1.4.2
 	github.com/spf13/cobra v1.3.0
+	github.com/spf13/pflag v1.0.5
 	github.com/spf13/viper v1.10.1
 	github.com/zmap/dns v1.1.45-zdns-0
 	github.com/zmap/go-iptree v0.0.0-20170831022036-1948b1097e25
@@ -24,7 +25,6 @@ require (
 	github.com/spf13/afero v1.6.0 // indirect
 	github.com/spf13/cast v1.4.1 // indirect
 	github.com/spf13/jwalterweatherman v1.1.0 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/subosito/gotenv v1.2.0 // indirect
 	golang.org/x/mod v0.5.0 // indirect
 	golang.org/x/net v0.0.0-20210813160813-60bc85c4be6d // indirect
